backend/pyproject.toml
[project]
name = "wb-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "dotenv>=0.9.9",
    "loguru>=0.7.3",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "ruff>=0.12.0",
]

[tool.ruff]
line-length = 79
target-version = "py312"
exclude = ["migrations", "__pycache__", "venv"]

[tool.ruff.lint]
select = [
  "E", "F", "I", "B", "UP", "C4", "TID"
]
